requirements.txt
transformers
sentence-transformers
pandas
python-dotenv
groq
torch
faiss-cpu
numpy
nltk
pymupdf  
